// src/app/api/models/update-user.ts
/* tslint:disable */
export class UpdateUser {
  applicationUserId?: null | string;
  email?: null | string;
  fullName?: null | string;
  password?: null | string;
  phoneNumber?: null | string;
  userName?: null | string;
}
export class UpdateUserWeb {
  applicationUserId?: null | string;
  email?: null | string;
  fullName?: null | string;
  image?: null | string;
  password?: null | string;
  phoneNumber?: null | string;
  userName?: null | string;
}